Unify personal and shared fittings under one URL, add publisher credit
Personal and published fittings lived at different route shapes
(/fitting/[characterId]/[fittingId] vs /fitting/shared/[sharedId]).
They now share one: /fitting/[fittingId], where the param is an opaque
token — a shared fit's uuid verbatim, or ${characterId}_${fittingId}
for a personal one. `_` never appears in a uuid's canonical form or in
a numeric fitting_id, so splitting on the last `_` unambiguously tells
the two apart (parseFittingRouteParam in fit.ts; personalFittingRoute/
sharedFittingRoute build the encoded hrefs everywhere a link is built:
the matrix, the publish/unpublish redirects, and the list_fittings MCP
tool's url field, replacing the fitting_id/character_id pair it used
to return).

The single page.tsx dispatches on the parsed kind to one of two async
components (PersonalFitting / SharedFitting) that carry the two
detail pages' previous bodies unchanged, so the merge is routing-only
— no behavior change to either besides the URL.

A shared fitting's page also gains a "Shared from <name>" badge with
the publisher's portrait, floating at the top right — the same
top-right-hugging-corner idiom the character list's avatars use.
Needs the publisher's EVE character id (not just their name), which
character_directory carries alongside the name already being looked
up there.
