+++ src/app/fitting/fit.ts
@@ -76,3 +76,32 @@ export const flagSortKey = (flag: string): number => {
   const index = FLAG_ORDER.findIndex((prefix) => flag.startsWith(prefix))
   return index === -1 ? FLAG_ORDER.length : index
 }
+
+// Every fitting on the site — personal or published — lives at one URL shape,
+// /fitting/[fittingId], so a link never has to know which kind it's pointing
+// at. The two sources don't share an id space, though: a shared_fitting row
+// has a globally unique uuid, but ESI numbers personal fittings per pilot (a
+// bigint that's only unique *within* one character — every character has a
+// fitting 1), so a bare fitting_id can't address one on its own.
+//
+// The route param is therefore an opaque token: a shared fit's id verbatim
+// (already a uuid), or `${characterId}_${fittingId}` for a personal one.
+// `_` never appears in a uuid's canonical hex-and-hyphen form or in a numeric
+// fitting_id, so splitting on the last `_` is unambiguous — a param with no
+// `_` is a shared uuid, one with an `_` is a personal fit's composite id.
+export const personalFittingRoute = (characterId: string, fittingId: number | string): string =>
+  `/fitting/${characterId}_${fittingId}`
+
+export const sharedFittingRoute = (sharedId: string): string => `/fitting/${sharedId}`
+
+export type FittingRouteParam =
+  { kind: 'personal'; characterId: string; fittingId: string } | { kind: 'shared'; sharedId: string }
+
+export const parseFittingRouteParam = (param: string): FittingRouteParam | null => {
+  const at = param.lastIndexOf('_')
+  if (at === -1) return { kind: 'shared', sharedId: param }
+  const characterId = param.slice(0, at)
+  const fittingId = param.slice(at + 1)
+  if (!characterId || !/^\d+$/.test(fittingId)) return null
+  return { kind: 'personal', characterId, fittingId }
+}
